Abstract TwoPaneLayoutCore
